+++ src/cineripr/__init__.py
@@ -2,4 +2,4 @@ from __future__ import annotations
 
 __all__ = ["__version__"]
 
-__version__ = "1.0.27"
+__version__ = "1.0.28"